+++ src/syntax/block/pretext.ts
@@ -8,20 +8,23 @@ import { squash } from '../inline/text';
 
 type SubParsers = [PlainTextParser];
 
-const syntax = /^```([0-9a-z]+)?[ \t]*\n/;
-const closer = /^```\s*(?:\n|$)/;
+const syntax = /^(`{3,})([0-9a-z]+)?[ \t]*\n[\s\S]*?\1/;
+const cache = new Map<string, RegExp>();
 
 export const pretext: PreTextParser = function (source: string): Result<HTMLPreElement, SubParsers> {
-  const [whole, lang] = source.match(syntax) || ['', ''];
+  const [whole, keyword, lang] = source.match(syntax) || ['', '', ''];
   if (!whole) return;
   const el = document.createElement('pre');
   if (lang) {
     void el.setAttribute('class', `lang-${lang.toLowerCase()}`);
   }
-  source = source.slice(whole.length);
+  source = source.slice(source.indexOf('\n') + 1);
+  if (!cache.has(keyword)) {
+    void cache.set(keyword, new RegExp(`^${keyword}\s*(?:\n|$)`));
+  }
   while (true) {
     const line = source.split('\n', 1)[0];
-    if (line.match(closer)) break;
+    if (line.match(cache.get(keyword)!)) break;
     void el.appendChild(squash((loop(compose<SubParsers, Text>([plaintext]))(line + '\n') || [[]])[0]));
     source = source.slice(line.length + 1);
     if (source === '') return;
@@ -29,5 +32,5 @@ export const pretext: PreTextParser = function (source: string): Result<HTMLPreE
   if (el.lastChild) {
     el.lastChild!.textContent = el.lastChild.textContent!.slice(0, -1);
   }
-  return consumeBlockEndEmptyLine<HTMLPreElement, SubParsers>([el], source.replace(closer, ''));
+  return consumeBlockEndEmptyLine<HTMLPreElement, SubParsers>([el], source.slice(keyword.length));
 }