+++ src/view/segment.test.ts
@@ -26,6 +26,7 @@ describe('Unit: binder/segment', () => {
       assert.deepStrictEqual(segment('```'), ['```']);
       assert.deepStrictEqual(segment('```\n```'), ['```\n```']);
       assert.deepStrictEqual(segment('```\n\n\n```'), ['```\n\n\n```']);
+      assert.deepStrictEqual(segment('````\n```\n````'), ['````\n```\n````']);
       assert.deepStrictEqual(segment('a\n\n\n```\n\n\n```'), ['a\n\n', '\n', '```\n\n\n```']);
       assert.deepStrictEqual(segment('a\n\n\n\n```\n\n\n```'), ['a\n\n', '\n\n', '```\n\n\n```']);
     });